requirements.txt
pycryptodome
pycaption
streamlink